requirements.txt
Django==1.6.2
PyYAML==3.10
South==0.8.4
django-angular==0.6.0
django-annoying==0.7.9
django-bootstrap-form==3.1
django-extensions==1.3.3
django-localflavor==1.0
django-recurrence==1.0.1
django-tastypie==0.11.0
docopt==0.4.0
mandrill==1.0.53
oauthlib==0.6.1
psycopg2==2.5.2
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
python-social-auth==0.1.21
pytz==2013.9
requests==2.2.1
requests-oauthlib==0.4.0
six==1.5.2
wsgiref==0.1.2
